+++ src/layout/section/highlightSection/HighlightSection.tsx
@@ -1,21 +1,19 @@
 import React from 'react'
 
 import { SectionContentSize } from '../../../layout/section/baseSection'
-import { SubHeader } from '../../../subHeader'
 import { StyledHighlightSection } from './HighlightSection.style'
 
 export type HighlightSectionProps = Readonly<{
   className?: string
   contentClassName?: string
   children: React.ReactNode
-  title?: string
 }>
 
 /**
  * A specialized section with an highlighting background color.
  */
 export const HighlightSection = (props: HighlightSectionProps) => {
-  const { className, children, contentClassName, title } = props
+  const { className, children, contentClassName } = props
   return (
     <StyledHighlightSection
       tagName="article"
@@ -23,7 +21,6 @@ export const HighlightSection = (props: HighlightSectionProps) => {
       contentClassName={contentClassName}
       contentSize={SectionContentSize.LARGE}
     >
-      {title && <SubHeader>{title}</SubHeader>}
       {children}
     </StyledHighlightSection>
   )